Refactored main loops into core.py
